% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_nflfastR_db}
\alias{update_nflfastR_db}
\title{Update the nflfastR Database}
\usage{
update_nflfastR_db(dbdir = "~/.db", force_rebuild = F)
}
\arguments{
\item{dbdir}{Character. The directory where the database is stored. Defaults to `~/.db`.}

\item{force_rebuild}{Logical. If `TRUE`, forces a rebuild of the database. Defaults to `FALSE`.}
}
\value{
No return value. Updates the nflfastR database in place.
}
\description{
This function is a wrapper that updates the `nflfastR` database by
calling `nflfastR::update_db()`.
}
\examples{
\dontrun{
update_nflfastR_db()  # Updates the database in the default location
update_nflfastR_db(dbdir = "data/nfl", force_rebuild = TRUE)  # Forces a rebuild
}
}
