% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_nflfastr_stats}
\alias{update_nflfastr_stats}
\title{Update tables related to `nflfastR::calculate_stats()`}
\usage{
update_nflfastr_stats(seasons)
}
\arguments{
\item{seasons}{numeric vector of seasons to specifiy.}
}
\description{
This function will update tables that are derived from `nflfastR::calculate_stast()`.
}
