% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_nflreadr_db}
\alias{update_nflreadr_db}
\title{main function to update tables that are related to `{nflreadr}`.}
\usage{
update_nflreadr_db(seasons = NULL)
}
\arguments{
\item{seasons}{numeric vector to indentify seasons to back filter for and backfill.
Note: this should typically only be used with `nflfastr::most_recent_season()` or maintain
the default value of `NULL`, as datasets have different years that started collecting. It is
easier to do a full refresh than handle all the different conditions.}
}
\description{
This function will update tables based on the ENV variable `DB_PATH` that is used in
`load_data()`.
}
