% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elt.R
\name{update_pbp}
\alias{update_pbp}
\title{Update the nflfastR Database}
\usage{
update_pbp(
  dbdir = "~/.db",
  dbname = "luna",
  force_rebuild = FALSE,
  tblname = DBI::Id(schema = "BASE", table = "NFLFASTR_PBP"),
  db_connection = DBI::dbConnect(duckdb::duckdb(), Sys.getenv("DB_PATH"))
)
}
\arguments{
\item{dbdir}{Character. The directory where the database is stored. Defaults to `~/.db`.}

\item{dbname}{Character. DB name, defaults to my dog `luna`}

\item{force_rebuild}{Logical. If `TRUE`, forces a rebuild of the database. Defaults to `FALSE`.}

\item{tblname}{name or DBI::Id(schema, table) of the table location}

\item{db_connection}{A `DBI::dbConnect()` object}
}
\value{
No return value. Updates the nflfastR database in place.
}
\description{
This function is a wrapper that updates the `nflfastR` database by
calling `nflfastR::update_db()`.
}
\examples{
\dontrun{
update_nflfastR_db()  # Updates the database in the default location
update_nflfastR_db(dbdir = "data/nfl", force_rebuild = TRUE)  # Forces a rebuild
}
}
